lark: fix golangci-lint findings across the new client
Cleans up every golangci-lint issue in the new internal/lark/ tree and the
internal/feishu/ rewrite so the repo passes `golangci-lint run ./...` at
0 issues:

- removed dead fields/methods (startOnce, readFinal, fakeServer.closeConn,
  serverSession.{conn,br,sendBinary,sendFragmented}, recordingHandler.errs)
- errcheck: resp.Body.Close under defer func(){} per the project convention
- bodyclose: capture+close the Dial response (idempotent; Dial already closes)
- gosec G115: bound-check decoded varints before int32/int (also a real
  robustness fix — a malformed 64-bit length can no longer wrap negative);
  bit-faithful int32→varint encode and tag/wire shifts tagged nolint with
  rationale (protocol-inherent, matches SDK pbbp2.pb.go)
- gosec G401/G505: SHA-1 in dial.go is MANDATED by RFC 6455 §1.3 (nolint)
- gosec G404: reconnect jitter uses math/rand/v2, not a security primitive
- noctx: tls.Dialer.DialContext replaces tls.DialWithDialer (context-aware)
- errorlint: errors.As for *APIError checks in tests
- canonicalheader / errname (fatalError) / unconvert / ineffassign / intrange
- recvcheck: dropped the mutating Headers.Add in favour of a free
  appendHeader helper, leaving Headers with consistent value receivers

Also fixes one pre-existing intrange finding in feishufront/progress_test.go
that was the last remaining issue project-wide.

+++ internal/feishu/bot_test.go
@@ -41,7 +41,7 @@ func (f *fakeClient) Send(context.Context, *lark.SendInput) (*lark.SendResult, e
 func (f *fakeClient) PatchMessage(_ context.Context, _, content string) error {
 	n := f.patchCalls.Add(1)
 	f.patchLast = content
-	if f.patchErrOnNth != 0 && int32(n) == f.patchErrOnNth {
+	if f.patchErrOnNth != 0 && n == f.patchErrOnNth {
 		return f.patchErr
 	}
 	if f.patchErrOnNth == 0 {
